requirements.txt
PyQt5==5.15.9
Pillow==9.0.1
fpdf==1.7.2
